+++ lib/kuby/prebundler/plugin.rb
@@ -3,13 +3,26 @@ require 'kuby'
 module Kuby
   module Prebundler
     class Plugin < ::Kuby::Plugin
+      attr_reader :config
+
+      def configure(&block)
+        config.instance_eval(&block) if block
+      end
+
       def after_configuration
         bundler_phase = docker.bundler_phase
-        docker.delete(:bundler_phase)
-        docker.insert(:bundler_phase, PrebundlerPhase.new(environment, bundler_phase), after: :package_phase)
+
+        docker.replace(
+          :bundler_phase,
+          PrebundlerPhase.new(environment, bundler_phase, config)
+        )
       end
 
       private
+
+      def after_initialize
+        @config = Config.new
+      end
 
       def docker
         environment.docker