+++ CHANGELOG.md
@@ -1,2 +1,6 @@
+## 0.2.0
+* Use kuby-core's `LayerStack#replace` instead of `#delete` and `#insert`, which leads to unexpected phase order.
+* Allow configuring the path to the config file.
+
 ## 0.1.0
 * Birthday!